Make up the fucking list
This takes in the file created by the `hyphenator` and then
writes a line to the console for every line in the input file.

It looks for the first hyphen and inserts `fucking-` after.
It then removes all of the remaining hyphens in the word.

We need some rules to handle things like if the last hyphened group
is an `-ing` then we don't want to add `fucking-` immediately before that.
It looks weird.

To use

```powershell
dotnet run ../data/hyphenated-adjectives.txt > ../data/fucking-adjectives.txt
```